Centralize the GPU state conversion functions
