+++ GPU/Common/ShaderId.h
@@ -135,3 +135,13 @@ struct ShaderID {
 		memcpy(d, &(src)[0], sizeof(d));
 	}
 };
+
+
+bool CanUseHardwareTransform(int prim);
+void ComputeVertexShaderID(ShaderID *id, u32 vertexType, bool useHWTransform);
+// Generates a compact string that describes the shader. Useful in a list to get an overview
+// of the current flora of shaders.
+std::string VertexShaderDesc(const ShaderID &id);
+
+void ComputeFragmentShaderID(ShaderID *id, uint32_t vertType);
+std::string FragmentShaderDesc(const ShaderID &id);
