+++ GPU/Directx9/PixelShaderGeneratorDX9.h
@@ -29,38 +29,6 @@ namespace DX9 {
 void ComputeFragmentShaderIDDX9(ShaderID *id);
 void GenerateFragmentShaderDX9(char *buffer);
 
-enum StencilValueType {
-	STENCIL_VALUE_UNIFORM,
-	STENCIL_VALUE_ZERO,
-	STENCIL_VALUE_ONE,
-	STENCIL_VALUE_KEEP,
-	STENCIL_VALUE_INVERT,
-	STENCIL_VALUE_INCR_4,
-	STENCIL_VALUE_INCR_8,
-	STENCIL_VALUE_DECR_4,
-	STENCIL_VALUE_DECR_8,
-};
-
-enum ReplaceAlphaType {
-	REPLACE_ALPHA_NO = 0,
-	REPLACE_ALPHA_YES = 1,
-	REPLACE_ALPHA_DUALSOURCE = 2,
-};
-
-enum ReplaceBlendType {
-	REPLACE_BLEND_NO,
-	REPLACE_BLEND_STANDARD,
-	REPLACE_BLEND_PRE_SRC,
-	REPLACE_BLEND_PRE_SRC_2X_ALPHA,
-	REPLACE_BLEND_2X_ALPHA,
-	REPLACE_BLEND_2X_SRC,
-	REPLACE_BLEND_COPY_FBO,
-};
-
-StencilValueType ReplaceAlphaWithStencilType();
-ReplaceAlphaType ReplaceAlphaWithStencil(ReplaceBlendType replaceBlend);
-ReplaceBlendType ReplaceBlendWithShader(bool allowShaderBlend);
-
 #define CONST_PS_TEXENV 0
 #define CONST_PS_ALPHACOLORREF 1
 #define CONST_PS_ALPHACOLORMASK 2
