+++ GPU/GPU.vcxproj.filters
@@ -192,13 +192,11 @@
     <ClInclude Include="GLES\GLStateCache.h">
       <Filter>GLES</Filter>
     </ClInclude>
-    <ClInclude Include="Common\GPUStateUtils.h">
-      <Filter>Common</Filter>
-    </ClInclude>
     <ClInclude Include="Common\ShaderCommon.h">
       <Filter>Common</Filter>
     </ClInclude>
     <ClInclude Include="Common\ShaderId.h" />
+    <ClInclude Include="Common\GPUStateUtils.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -379,5 +377,6 @@
       <Filter>GLES</Filter>
     </ClCompile>
     <ClCompile Include="Common\ShaderId.cpp" />
+    <ClCompile Include="Common\GPUStateUtils.cpp" />
   </ItemGroup>
 </Project>
